requirements/base.txt
asgiref
cfgv
distlib
Django
django-ckeditor
django-cors-headers
django-filter
django-jazzmin
django-js-asset
django-phonenumber-field
djangorestframework
drf-yasg
filelock
identify
inflection
nodeenv
packaging
Pillow
platformdirs
pre-commit
psycopg2-binary
python-environ
pytz
PyYAML
sqlparse
typing_extensions
uritemplate
virtualenv
django-filter
opencv-python

